requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
distlib==0.3.6
dnspython==2.3.0
editables==0.4
email-validator==2.0.0.post2
fastapi==0.99.1
fastapi-cache2==0.2.1
fastapi-users==12.0.0
fastapi-users-db-sqlalchemy==6.0.0
filelock==3.12.2
flower==2.0.0
frozenlist==1.3.3
gevent==23.7.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hatchling==1.18.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
humanize==4.7.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==24.2.0
kombu==5.3.1
makefun==1.15.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
orjson==3.9.1
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
platformdirs==3.8.1
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyperclip==1.8.2
pytest==7.4.0
pytest-anyio==0.0.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.6.0
rich==13.4.2
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli_w==1.0.0
tomlkit==0.11.8
tornado==6.3.2
trove-classifiers==2023.7.6
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.3
userpath==1.8.0
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.23.1
watchfiles==0.19.0
wcwidth==0.2.6
websocket==0.2.1
websockets==11.0.3
yarl==1.9.2
zipp==3.16.1
zope.event==5.0
zope.interface==6.0
